requirements.txt
feedparser
docopt
python-dateutil